Extract branding texts model as uiState
