+++ WordPress/src/main/java/org/wordpress/android/ui/jetpackoverlay/JetpackFeatureRemovalBrandingUtil.kt
@@ -2,12 +2,12 @@ package org.wordpress.android.ui.jetpackoverlay
 
 import org.wordpress.android.R
 import org.wordpress.android.models.JetpackPoweredScreen
-import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalBrandingUtil.Interval.Days
-import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalBrandingUtil.Interval.Indeterminate
-import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalBrandingUtil.Interval.Passed
-import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalBrandingUtil.Interval.Pluralisable
-import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalBrandingUtil.Interval.Soon
-import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalBrandingUtil.Interval.Weeks
+import org.wordpress.android.ui.jetpackoverlay.JetpackBrandingUiState.Days
+import org.wordpress.android.ui.jetpackoverlay.JetpackBrandingUiState.Indeterminate
+import org.wordpress.android.ui.jetpackoverlay.JetpackBrandingUiState.Passed
+import org.wordpress.android.ui.jetpackoverlay.JetpackBrandingUiState.Pluralisable
+import org.wordpress.android.ui.jetpackoverlay.JetpackBrandingUiState.Soon
+import org.wordpress.android.ui.jetpackoverlay.JetpackBrandingUiState.Weeks
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalPhase.PhaseFour
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalPhase.PhaseOne
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalPhase.PhaseThree
@@ -20,9 +20,6 @@ import org.wordpress.android.util.DateTimeUtilsWrapper
 import org.wordpress.android.util.config.JPDeadlineConfig
 import java.time.LocalDate
 import java.time.ZoneId
-import java.time.temporal.ChronoUnit.DAYS
-import java.time.temporal.ChronoUnit.MONTHS
-import java.time.temporal.ChronoUnit.WEEKS
 import java.util.Date
 import javax.inject.Inject
 
@@ -59,8 +56,8 @@ class JetpackFeatureRemovalBrandingUtil @Inject constructor(
             PhaseTwo -> UiStringRes(R.string.wp_jetpack_powered_phase_2)
             PhaseThree -> (screen as? JetpackPoweredScreen.WithDynamicText)?.let { screenWithDynamicText ->
                 getDynamicBrandingForScreen(screenWithDynamicText)
-            } ?: UiStringRes(Interval.RES_JP_POWERED)
-            else -> UiStringRes(Interval.RES_JP_POWERED)
+            } ?: UiStringRes(JetpackBrandingUiState.RES_JP_POWERED)
+            else -> UiStringRes(JetpackBrandingUiState.RES_JP_POWERED)
         }
     }
 
@@ -78,11 +75,11 @@ class JetpackFeatureRemovalBrandingUtil @Inject constructor(
         deadline: LocalDate,
     ): UiString {
         val today = dateTimeUtilsWrapper.getTodaysDate().toLocalDate()
-        return when (val interval = Interval.between(today, deadline)) {
+        return when (val interval = JetpackBrandingUiState.between(today, deadline)) {
             is Indeterminate -> getPhaseThreeMovingSoonBranding(screen)
             is Weeks -> getMovingInUiString(screen, getQuantityUiString(interval))
             is Days -> getMovingInUiString(screen, getQuantityUiString(interval))
-            is Passed -> UiStringRes(Interval.RES_JP_POWERED)
+            is Passed -> UiStringRes(JetpackBrandingUiState.RES_JP_POWERED)
         }
     }
 
@@ -133,50 +130,4 @@ class JetpackFeatureRemovalBrandingUtil @Inject constructor(
         }
     }
 
-    sealed class Interval {
-        sealed interface Indeterminate
-        sealed interface Pluralisable {
-            val number: Long
-            val oneRes: Int
-            val otherRes: Int
-
-            companion object {
-                const val RES_ARE_MOVING_IN = R.string.wp_jetpack_powered_phase_3_feature_are_moving_in
-                const val RES_IS_MOVING_IN = R.string.wp_jetpack_powered_phase_3_feature_is_moving_in
-            }
-        }
-
-        object Soon : Interval(), Indeterminate {
-            const val RES_ARE_MOVING_SOON = R.string.wp_jetpack_powered_phase_3_feature_are_moving_soon
-            const val RES_IS_MOVING_SOON = R.string.wp_jetpack_powered_phase_3_feature_is_moving_soon
-        }
-
-        data class Weeks(override val number: Long) : Interval(), Pluralisable {
-            override val oneRes = R.string.weeks_quantity_one
-            override val otherRes = R.string.weeks_quantity_other
-        }
-
-        data class Days(override val number: Long) : Interval(), Pluralisable {
-            override val oneRes = R.string.days_quantity_one
-            override val otherRes = R.string.days_quantity_other
-        }
-
-        object Unknown : Interval(), Indeterminate
-        object Passed : Interval()
-
-        companion object {
-            const val RES_JP_POWERED = R.string.wp_jetpack_powered
-
-            fun between(startDate: LocalDate, endDate: LocalDate): Interval {
-                val days = DAYS.between(startDate, endDate)
-                return when {
-                    days > MONTHS.duration.toDays() -> Soon
-                    days >= WEEKS.duration.toDays() -> Weeks(WEEKS.between(startDate, endDate))
-                    days >= 1 || startDate == endDate -> Days(days.coerceAtLeast(1))
-                    days < 0 -> Passed
-                    else -> Unknown
-                }
-            }
-        }
-    }
 }
